+++ config.py
@@ -1,0 +1,4 @@
+import os
+
+SECRET_KEY = 'alura'
+UPLOAD_PATH = os.path.dirname(os.path.abspath(__file__)) + '/uploads'
