+++ jogoteca.py
@@ -1,131 +1,10 @@
 import os
-from flask import (Flask, flash, redirect, render_template, request, session,
-                   url_for, send_from_directory)
-from App.models.jogo import Jogo
-from persistence.database.dao import JogoDao, UsuarioDao
-from persistence.database.database import Session
-import time
+from flask import Flask
 
 app = Flask(__name__)
-app.secret_key = 'alura'
-app.config['UPLOAD_PATH'] = os.path.dirname(
-    os.path.abspath(__file__)) + '/uploads'
+app.config.from_pyfile('config.py')
 
-jogo_dao = JogoDao(Session)
-usuario_dao = UsuarioDao(Session)
+from views import *
 
-
-@app.route('/')
-def index():
-    lista = jogo_dao.listar()
-    return render_template('lista.html', titulo='Jogos', jogos=lista)
-
-
-@app.route('/novo')
-def novo():
-    if 'usuario_logado' not in session or session['usuario_logado'] == None:
-        return redirect(url_for('login', proxima=url_for('novo')))
-    return render_template('novo.html', titulo='Novo Jogo')
-
-
-@app.route('/criar', methods=['POST', ])
-def criar():
-    nome = request.form['nome']
-    categoria = request.form['categoria']
-    console = request.form['console']
-
-    jogo = Jogo(nome, categoria, console)
-    jogo = jogo_dao.salvar(jogo)
-
-    arquivo = request.files['arquivo']
-    upload_path = app.config['UPLOAD_PATH']
-
-    ultimo_id = jogo_dao.listar()[-1]
-
-    timestamp = time.time()
-
-    arquivo.save(
-        f"{upload_path}/capa-{ultimo_id.id}-{timestamp}.jpg")
-
-    return redirect(url_for('index'))
-
-
-@app.route('/editar/<int:id>')
-def editar(id):
-    if 'usuario_logado' not in session or session['usuario_logado'] == None:
-        return redirect(url_for('login', proxima=url_for('editar')))
-    jogo = jogo_dao.busca_por_id(id)
-    nome_imagem = recupera_imagem(id)
-    return render_template('editar.html', titulo='Editando Jogo', jogo=jogo, capa_jogo=nome_imagem)
-
-
-@app.route('/atualizar', methods=['POST', ])
-def atualizar():
-    nome = request.form['nome']
-    categoria = request.form['categoria']
-    console = request.form['console']
-
-    jogo = Jogo(nome, categoria, console, id=request.form['id'])
-    jogo = jogo_dao.salvar(jogo)
-
-    arquivo = request.files['arquivo']
-    upload_path = app.config['UPLOAD_PATH']
-    
-    timestamp = time.time()
-    deleta_arquivo(jogo.id)
-    arquivo.save(f'{upload_path}/capa-{jogo.id}-{timestamp}.jpg')
-
-    return redirect(url_for('index'))
-
-
-@app.route('/deletar/<int:id>')
-def deletar(id):
-    jogo_dao.deletar(id)
-    flash('O jogo foi removido com sucesso!')
-    return redirect(url_for('index'))
-
-
-@app.route('/login')
-def login():
-    proxima = request.args.get('proxima')
-    return render_template('login.html', proxima=proxima)
-
-
-@app.route('/autenticar', methods=['POST', ])
-def autenticar():
-    usuario = usuario_dao.buscar_por_id(request.form['usuario'])
-    if usuario:
-        if request.form['senha'] == usuario.senha:
-            session['usuario_logado'] = usuario.nome
-            flash(usuario.nome +
-                  ' logado(a) com sucesso!')
-            proxima_pagina = request.form['proxima']
-            return redirect(proxima_pagina)
-    else:
-        flash('Usuário não logado!')
-        return redirect(url_for('login'))
-
-
-@app.route('/logout')
-def logout():
-    session['usuario_logado'] = None
-    flash('Logout efetuado com sucesso!')
-    return redirect(url_for('index'))
-
-
-@app.route('/uploads/<nome_arquivo>')
-def imagem(nome_arquivo):
-    return send_from_directory('uploads', nome_arquivo)
-
-
-def recupera_imagem(id):
-    for nome_arquivo in os.listdir(app.config['UPLOAD_PATH']):
-        if f'capa-{id}' in nome_arquivo:
-            return nome_arquivo
-        
-def deleta_arquivo(id):
-    arquivo = recupera_imagem(id)
-    os.remove(os.path.join(app.config['UPLOAD_PATH'], arquivo))
-
-
-app.run(host='0.0.0.0', port=8080, debug=True)
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=8080, debug=True)
