+++ views.py
@@ -1,0 +1,114 @@
+import time
+from flask import (Flask, flash, redirect, render_template, request,
+                   send_from_directory, session, url_for)
+from App.models.jogo import Jogo
+from helpers import deleta_arquivo, recupera_imagem
+from jogoteca import app
+from persistence.database.dao import JogoDao, UsuarioDao
+from persistence.database.database import Session
+
+jogo_dao = JogoDao(Session)
+usuario_dao = UsuarioDao(Session)
+
+
+@app.route('/')
+def index():
+    lista = jogo_dao.listar()
+    return render_template('lista.html', titulo='Jogos', jogos=lista)
+
+
+@app.route('/novo')
+def novo():
+    if 'usuario_logado' not in session or session['usuario_logado'] == None:
+        return redirect(url_for('login', proxima=url_for('novo')))
+    return render_template('novo.html', titulo='Novo Jogo')
+
+
+@app.route('/criar', methods=['POST', ])
+def criar():
+    nome = request.form['nome']
+    categoria = request.form['categoria']
+    console = request.form['console']
+
+    jogo = Jogo(nome, categoria, console)
+    jogo = jogo_dao.salvar(jogo)
+
+    arquivo = request.files['arquivo']
+    upload_path = app.config['UPLOAD_PATH']
+
+    ultimo_id = jogo_dao.listar()[-1]
+
+    timestamp = time.time()
+
+    arquivo.save(
+        f"{upload_path}/capa-{ultimo_id.id}-{timestamp}.jpg")
+
+    return redirect(url_for('index'))
+
+
+@app.route('/editar/<int:id>')
+def editar(id):
+    if 'usuario_logado' not in session or session['usuario_logado'] == None:
+        return redirect(url_for('login', proxima=url_for('editar')))
+    jogo = jogo_dao.busca_por_id(id)
+    nome_imagem = recupera_imagem(id)
+    return render_template('editar.html', titulo='Editando Jogo', jogo=jogo, capa_jogo=nome_imagem)
+
+
+@app.route('/atualizar', methods=['POST', ])
+def atualizar():
+    nome = request.form['nome']
+    categoria = request.form['categoria']
+    console = request.form['console']
+
+    jogo = Jogo(nome, categoria, console, id=request.form['id'])
+    jogo = jogo_dao.salvar(jogo)
+
+    arquivo = request.files['arquivo']
+    upload_path = app.config['UPLOAD_PATH']
+
+    timestamp = time.time()
+    deleta_arquivo(jogo.id)
+    arquivo.save(f'{upload_path}/capa-{jogo.id}-{timestamp}.jpg')
+
+    return redirect(url_for('index'))
+
+
+@app.route('/deletar/<int:id>')
+def deletar(id):
+    jogo_dao.deletar(id)
+    flash('O jogo foi removido com sucesso!')
+    return redirect(url_for('index'))
+
+
+@app.route('/login')
+def login():
+    proxima = request.args.get('proxima')
+    return render_template('login.html', proxima=proxima)
+
+
+@app.route('/autenticar', methods=['POST', ])
+def autenticar():
+    usuario = usuario_dao.buscar_por_id(request.form['usuario'])
+    if usuario:
+        if request.form['senha'] == usuario.senha:
+            session['usuario_logado'] = usuario.nome
+            flash(usuario.nome +
+                  ' logado(a) com sucesso!')
+            proxima_pagina = request.form['proxima']
+            return redirect(proxima_pagina)
+    else:
+        flash('Usuário não logado!')
+        return redirect(url_for('login'))
+
+
+@app.route('/logout')
+def logout():
+    session['usuario_logado'] = None
+    flash('Logout efetuado com sucesso!')
+    return redirect(url_for('index'))
+
+
+@app.route('/uploads/<nome_arquivo>')
+def imagem(nome_arquivo):
+    return send_from_directory('uploads', nome_arquivo)
